God, help me, please!

+++ tasks/mpi/kolodkin_g_sentence_count/perf_tests/main.cpp
@@ -13,7 +13,7 @@ TEST(mpi_example_perf_test, test_pipeline_run) {
       "verifwriefnifnil!?vfnjklererjerjkerg...vrhklererffwjklfwefwejo!vefnklvevef?wfnkrkflwewefkl!vfnklvfklevf?"
       "vrrnervevrnvreiev!";
   std::vector<char> global_str;
-  for (int i = 0; i < str.length(); i++) {
+  for (unsigned long int i = 0; i < str.length(); i++) {
     global_str.push_back(str[i]);
   }
   std::vector<int32_t> global_sum(1, 0);
@@ -56,7 +56,7 @@ TEST(mpi_example_perf_test, test_task_run) {
       "Na krayu dorogi stoyal dub! Eto byl ogromnuy, v dva obhvata dub. Knyaz Andrey podosel k dubu! Boze prabiy! Kak "
       "tebya zovut? Ya dub! A ya knyaz Andrey! Zdorovo! Poka-poka, dub! Poka, Andrey!";
   std::vector<char> global_str;
-  for (int i = 0; i < str.length(); i++) {
+  for (unsigned long int i = 0; i < str.length(); i++) {
     global_str.push_back(str[i]);
   }
   std::vector<int32_t> global_sum(1, 0);
