+++ tasks/seq/kolodkin_g_sentence_count/func_tests/main.cpp
@@ -9,7 +9,7 @@ TEST(Sequential, Test_two_sentences) {
   // Create data
   std::string in = "Hello! My name is Grisha!";
   std::vector<char> global_str;
-  for (int i = 0; i < in.length(); i++) {
+  for (unsigned long int i = 0; i < in.length(); i++) {
     global_str.push_back(in[i]);
   }
   std::vector<int> out(1, 0);
@@ -58,7 +58,7 @@ TEST(Sequential, Test_sentences_with_special_symbols_in_end_of_sentence) {
   std::string in =
       "Hello!My name is Grisha! I have two pets: cat,dog,parrot. What is your name?! How are you!? Well...";
   std::vector<char> global_str;
-  for (int i = 0; i < in.length(); i++) {
+  for (unsigned long int i = 0; i < in.length(); i++) {
     global_str.push_back(in[i]);
   }
   std::vector<int> out(1, 0);
@@ -83,7 +83,7 @@ TEST(Sequential, Test_sentences_with_double_symbols) {
   std::string in =
       "Hello!! My name is Grisha!! I have two pets: cat,dog,parrot. What is your name?! How are you!? Well...";
   std::vector<char> global_str;
-  for (int i = 0; i < in.length(); i++) {
+  for (unsigned long int i = 0; i < in.length(); i++) {
     global_str.push_back(in[i]);
   }
   std::vector<int> out(1, 0);
@@ -110,7 +110,7 @@ TEST(Sequential, Big_text) {
       "nebeseh i na zemle. Hleb nas nasyshnii dazd nam dnes, i ostavi nam dolgi nasha. Yakozhe i my ostavlyaem "
       "dolznikom nashim! I ne vvedi nas vo iskushenie, no izbavi nas ot lukavogo... Amin!";
   std::vector<char> global_str;
-  for (int i = 0; i < in.length(); i++) {
+  for (unsigned long int i = 0; i < in.length(); i++) {
     global_str.push_back(in[i]);
   }
   std::vector<int> out(1, 0);
