+++ tasks/seq/kolodkin_g_sentence_count/perf_tests/main.cpp
@@ -12,7 +12,7 @@ TEST(sequential_example_perf_test, test_pipeline_run) {
       "verifwriefnifnil!?vfnjklererjerjkerg...vrhklererffwjklfwefwejo!vefnklvevef?wfnkrkflwewefkl!vfnklvfklevf?"
       "vrrnervevrnvreiev!";
   std::vector<char> global_str;
-  for (int i = 0; i < str.length(); i++) {
+  for (unsigned long int i = 0; i < str.length(); i++) {
     global_str.push_back(str[i]);
   }
   std::vector<int> out(1, 0);
@@ -51,7 +51,7 @@ TEST(sequential_example_perf_test, test_task_run) {
   // Create data
   std::string str = "Hello! My name is Grisha! Good morning! How are you!";
   std::vector<char> global_str;
-  for (int i = 0; i < str.length(); i++) {
+  for (unsigned long int i = 0; i < str.length(); i++) {
     global_str.push_back(str[i]);
   }
   std::vector<int> out(1, 0);
